requirements.txt
fastapi
uvicorn[standard]
pyjwt
pydantic
sqlmodel
'pydantic[email]'
